requirements.txt
# --- Backend Dependencies ---
Flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
google-generativeai==0.3.2
pymysql==1.1.0
SQLAlchemy==2.0.23

# --- Frontend (Streamlit App) Dependencies ---
streamlit==1.35.0
requests==2.31.0
pandas==2.2.2
plotly==5.21.0
